affichage de la valeur du progress en int et non en double

+++ YggToolsPortable/Classes/TorrentInformation.cs
@@ -10,7 +10,7 @@ namespace YggToolsPortable.Classes
     class TorrentInformation 
     {
         public string Title { get; set; }
-        public double Completion { get; set; }
+        public int Completion { get; set; }
         public String DownSpeed { get; set; }
         public String UpSpeed { get; set; }
     }
